backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
asyncio
streamlit>=1.32.0
boto3>=1.34.0
botocore>=1.34.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
